+++ comic_dl/honcho.py
@@ -18,7 +18,7 @@ import globalFunctions
 
 
 class Honcho(object):
-    def checker(self, comic_url, download_directory, **kwargs):
+    def checker(self, comic_url, download_directory, chapter_range, **kwargs):
         
         user_name = kwargs.get("username")
         password = kwargs.get("password")
@@ -37,7 +37,7 @@ class Honcho(object):
 
         if domain in fool_slide:
             foolSlide.FoolSlide(manga_url=comic_url, logger=logging, current_directory=current_directory,
-                                sorting_order=sorting, log_flag=log_flag, download_directory=download_directory)
+                                sorting_order=sorting, log_flag=log_flag, download_directory=download_directory, chapter_range=chapter_range)
             return 0
         elif domain in ["www.readcomiconline.to", "readcomiconline.to"]:
             readcomicOnlineto.ReadComicOnlineTo(manga_url=comic_url, logger=logging,
@@ -45,27 +45,27 @@ class Honcho(object):
             return 0
         elif domain in ["www.comic.naver.com", "comic.naver.com"]:
             comicNaver.ComicNaver(manga_url=comic_url, logger=logging, current_directory=current_directory,
-                                  sorting_order=sorting, log_flag=log_flag, download_directory=download_directory)
+                                  sorting_order=sorting, log_flag=log_flag, download_directory=download_directory, chapter_range=chapter_range)
             return 0
         elif domain in ["www.mangahere.co", "mangahere.co"]:
             mangaHere.MangaHere(manga_url=comic_url, logger=logging, current_directory=current_directory,
-                                sorting_order=sorting, log_flag=log_flag, download_directory=download_directory)
+                                sorting_order=sorting, log_flag=log_flag, download_directory=download_directory, chapter_range=chapter_range)
             return 0
         elif domain in ["www.raw.senmanga.com", "raw.senmanga.com"]:
             rawSenManga.RawSenaManga(manga_url=comic_url, logger=logging, current_directory=current_directory,
-                                     sorting_order=sorting, log_flag=log_flag, download_directory=download_directory)
+                                     sorting_order=sorting, log_flag=log_flag, download_directory=download_directory, chapter_range=chapter_range)
             return 0
         elif domain in ["www.mangafox.me", "mangafox.me"]:
             mangaFox.MangaFox(manga_url=comic_url, logger=logging, current_directory=current_directory,
-                              sorting_order=sorting, log_flag=log_flag, download_directory=download_directory)
+                              sorting_order=sorting, log_flag=log_flag, download_directory=download_directory, chapter_range=chapter_range)
             return 0
         elif domain in ["www.omgbeaupeep.com", "omgbeaupeep.com"]:
             omgBeauPeep.OmgBeauPeep(manga_url=comic_url, logger=logging, current_directory=current_directory,
-                                    sorting_order=sorting, log_flag=log_flag, download_directory=download_directory)
+                                    sorting_order=sorting, log_flag=log_flag, download_directory=download_directory, chapter_range=chapter_range)
             return 0
         elif domain in ["www.ac.qq.com", "ac.qq.com"]:
             acQQ.AcQq(manga_url=comic_url, logger=logging, current_directory=current_directory,
-                                    sorting_order=sorting, log_flag=log_flag, download_directory=download_directory)
+                                    sorting_order=sorting, log_flag=log_flag, download_directory=download_directory, chapter_range=chapter_range)
             return 0
         elif domain in ["www.kissmanga.com", "kissmanga.com"]:
             # kissManga.KissManga(manga_url = comic_url, logger = logging, current_directory = current_directory, sorting_order = sorting)