+++ comic_dl/sites/foolSlide.py
@@ -9,7 +9,7 @@ import logging
 
 
 class FoolSlide(object):
-    def __init__(self, manga_url, download_directory, **kwargs):
+    def __init__(self, manga_url, download_directory, chapter_range, **kwargs):
 
         current_directory = kwargs.get("current_directory")
         self.logging = kwargs.get("log_flag")
@@ -18,7 +18,7 @@ class FoolSlide(object):
         self.manga_name = self.name_cleaner(manga_url)
 
         if "/reader/series/" in manga_url:
-            self.full_manga(manga_url=manga_url, comic_name=self.manga_name, sorting=self.sorting, download_directory=download_directory)
+            self.full_manga(manga_url=manga_url, comic_name=self.manga_name, sorting=self.sorting, download_directory=download_directory, chapter_range=chapter_range)
         elif "/reader/read/" in manga_url:
             self.single_chapter(manga_url, self.manga_name, download_directory)
 
@@ -69,7 +69,7 @@ class FoolSlide(object):
 
         return anime_name
 
-    def full_manga(self, manga_url, comic_name, sorting, download_directory, **kwargs):
+    def full_manga(self, manga_url, comic_name, sorting, download_directory, chapter_range, **kwargs):
         source, cookies = globalFunctions.GlobalFunctions().page_downloader(manga_url=manga_url)
         # print(source)
         chapter_text = source.findAll('div', {'class': 'title'})
@@ -81,6 +81,17 @@ class FoolSlide(object):
                 url = a['href']
                 all_links.append(url)
         logging.debug("All Links : %s" % all_links)
+
+        # Uh, so the logic is that remove all the unnecessary chapters beforehand and then pass the list for further operations.
+        if chapter_range != "All":
+            # -1 to shift the episode number accordingly to the INDEX of it. List starts from 0 xD!
+            starting = int(str(chapter_range).split("-")[0]) - 1
+            ending = int(str(chapter_range).split("-")[1])
+            indexes = [x for x in range(starting, ending)]
+            # [::-1] in sub_list in beginning to start this from the 1st episode and at the last, it is to reverse the list again, becasue I'm reverting it again at the end.
+            all_links = [all_links[::-1][x] for x in indexes][::-1]
+        else:
+            all_links = all_links
 
         if str(sorting).lower() in ['new', 'desc', 'descending', 'latest']:
             for chap_link in all_links: