+++ comic_dl/sites/omgBeauPeep.py
@@ -8,13 +8,20 @@ import logging
 
 
 class OmgBeauPeep(object):
-    def __init__(self, manga_url, download_directory, **kwargs):
+    def __init__(self, manga_url, download_directory, chapter_range, **kwargs):
 
         current_directory = kwargs.get("current_directory")
         self.logging = kwargs.get("log_flag")
         self.sorting = kwargs.get("sorting_order")
         self.comic_name = self.name_cleaner(manga_url)
-        self.single_chapter(manga_url, self.comic_name, download_directory)
+
+        # Since this website doesn't seem to have a dedicated page for series, I need to make a function just for the sake of range, huh. *sigh*
+        # You need to pass the ABSOLUTE chapter number for this website though. Otherwise, it'll generate wrong links to chapter and won't download anything.
+        # Please spare me such hard work.
+        if chapter_range != "All":
+            self.range_maker(manga_url=manga_url, chapter_range=chapter_range, download_directory=download_directory)
+        else:
+            self.single_chapter(manga_url, self.comic_name, download_directory)
 
     def name_cleaner(self, url):
         initial_name = str(url).split("/")[4].strip()
@@ -49,3 +56,12 @@ class OmgBeauPeep(object):
             logging.debug("Chapter Url : %s" % chapter_url)
             logging.debug("Image Link : %s" % image_link)
             globalFunctions.GlobalFunctions().downloader(image_link, file_name, chapter_url, directory_path, log_flag=self.logging)
+
+    def range_maker(self, manga_url, chapter_range, download_directory):
+        starting = int(str(chapter_range).split("-")[0])
+        ending = int(str(chapter_range).split("-")[1])
+        for comic_chapter in range(starting, ending):
+            chapter_number = str(manga_url).split("/")[5].strip()
+            # Dirty Hack, 'cause I'm a HackerMan! (Dumb Joke)
+            new_url = str(manga_url).replace(str(chapter_number), str(comic_chapter))
+            self.single_chapter(new_url, self.comic_name, download_directory)