+++ comic_dl/sites/rawSenManga.py
@@ -8,7 +8,7 @@ import logging
 
 
 class RawSenaManga(object):
-    def __init__(self, manga_url, download_directory, **kwargs):
+    def __init__(self, manga_url, download_directory, chapter_range, **kwargs):
         current_directory = kwargs.get("current_directory")
         self.logging = kwargs.get("log_flag")
         self.sorting = kwargs.get("sorting_order")
@@ -16,7 +16,7 @@ class RawSenaManga(object):
         url_split = str(manga_url).split("/")
 
         if len(url_split) is 5:
-            self.full_series(comic_url=manga_url, comic_name=self.comic_name, sorting=self.sorting, download_directory=download_directory)
+            self.full_series(comic_url=manga_url, comic_name=self.comic_name, sorting=self.sorting, download_directory=download_directory, chapter_range=chapter_range)
         else:
             self.single_chapter(manga_url, self.comic_name, download_directory)
 
@@ -62,7 +62,7 @@ class RawSenaManga(object):
                                                              cookies=cookies_main, log_flag=self.logging)
         return 0
 
-    def full_series(self, comic_url, comic_name, sorting, download_directory, **kwargs):
+    def full_series(self, comic_url, comic_name, sorting, download_directory, chapter_range, **kwargs):
         series_name_raw = str(comic_url).split("/")[3].strip()
         source, cookies = globalFunctions.GlobalFunctions().page_downloader(manga_url=comic_url)
         # a href="/Flying-Witch-Ishizuka-Chihiro/34/1"
@@ -70,6 +70,17 @@ class RawSenaManga(object):
 
         all_links = list(re.findall(link_regex, str(source)))
         logging.debug("All Links : %s" % all_links)
+
+        # Uh, so the logic is that remove all the unnecessary chapters beforehand and then pass the list for further operations.
+        if chapter_range != "All":
+            # -1 to shift the episode number accordingly to the INDEX of it. List starts from 0 xD!
+            starting = int(str(chapter_range).split("-")[0]) - 1
+            ending = int(str(chapter_range).split("-")[1])
+            indexes = [x for x in range(starting, ending)]
+            # [::-1] in sub_list in beginning to start this from the 1st episode and at the last, it is to reverse the list again, becasue I'm reverting it again at the end.
+            all_links = [all_links[::-1][x] for x in indexes][::-1]
+        else:
+            all_links = all_links
 
         if str(sorting).lower() in ['new', 'desc', 'descending', 'latest']:
             for link in all_links: